prov/tcp2: Move pep calls into their own source file
Simple code movement and restructuring.
