+++ libfabric.vcxproj
@@ -723,16 +723,17 @@
     <ClCompile Include="prov\tcp\src\tcpx_progress.c" />
     <ClCompile Include="prov\tcp2\src\tcp2_attr.c" />
     <ClCompile Include="prov\tcp2\src\tcp2_cm.c" />
-    <ClCompile Include="prov\tcp2\src\tcp2_srx.c" />
     <ClCompile Include="prov\tcp2\src\tcp2_cq.c" />
     <ClCompile Include="prov\tcp2\src\tcp2_domain.c" />
-    <ClCompile Include="prov\tcp2\src\tcp2_rma.c" />
-    <ClCompile Include="prov\tcp2\src\tcp2_msg.c" />
     <ClCompile Include="prov\tcp2\src\tcp2_ep.c" />
-    <ClCompile Include="prov\tcp2\src\tcp2_fabric.c" />
     <ClCompile Include="prov\tcp2\src\tcp2_eq.c" />
+    <ClCompile Include="prov\tcp2\src\tcp2_fabric.c" />
     <ClCompile Include="prov\tcp2\src\tcp2_init.c" />
+    <ClCompile Include="prov\tcp2\src\tcp2_msg.c" />
+    <ClCompile Include="prov\tcp2\src\tcp2_pep.c" />
     <ClCompile Include="prov\tcp2\src\tcp2_progress.c" />
+    <ClCompile Include="prov\tcp2\src\tcp2_rma.c" />
+    <ClCompile Include="prov\tcp2\src\tcp2_srx.c" />
     <ClCompile Include="prov\udp\src\udpx_attr.c" />
     <ClCompile Include="prov\udp\src\udpx_cq.c" />
     <ClCompile Include="prov\udp\src\udpx_domain.c" />
