+++ prov/tcp2/Makefile.include
@@ -7,6 +7,7 @@ _tcp2_files = \
 	prov/tcp2/src/tcp2_rma.c	\
 	prov/tcp2/src/tcp2_msg.c	\
 	prov/tcp2/src/tcp2_ep.c		\
+	prov/tcp2/src/tcp2_pep.c	\
 	prov/tcp2/src/tcp2_srx.c	\
 	prov/tcp2/src/tcp2_cq.c		\
 	prov/tcp2/src/tcp2_eq.c		\
